Features Settings as a screen groug; Removes workaround parent stacking issue on auto dismiss on useDismissCurrentRoute.
