+++ cardstack/src/navigation/screenGroups/index.ts
@@ -1,2 +1,3 @@
 export * from './profile';
 export * from './backup';
+export * from './settings';
