+++ cardstack/src/navigation/tabBarNavigator.tsx
@@ -28,11 +28,14 @@ import { loadAddress } from '@rainbow-me/model/wallet';
 import ChangeWalletSheet from '@rainbow-me/screens/ChangeWalletSheet';
 import ModalScreen from '@rainbow-me/screens/ModalScreen';
 import PinAuthenticationScreen from '@rainbow-me/screens/PinAuthenticationScreen';
-import SettingsModal from '@rainbow-me/screens/SettingsModal';
 
 import { createCustomStackNavigator } from './customNavigator';
 import { useCardstackMainScreens } from './hooks';
-import { ProfileScreenGroup, BackupScreenGroup } from './screenGroups';
+import {
+  ProfileScreenGroup,
+  BackupScreenGroup,
+  SettingsGroup,
+} from './screenGroups';
 
 import {
   dismissAndroidKeyboardOnClose,
@@ -232,7 +235,7 @@ export const StackNavigator = () => {
       {SharedScreens({ navigationKey })}
       {ProfileScreenGroup({ Stack })}
       {BackupScreenGroup({ Stack, navigationKey })}
-      {SettingsModal({ Stack })}
+      {SettingsGroup({ Stack })}
     </Stack.Navigator>
   );
 };
